requirements.txt
# https://pypi.org/project/flask
Flask==3.0.2

# https://pypi.org/project/gunicorn
gunicorn==21.2.0

#https://pypi.org/project/google-generativeai/
google-generativeai==0.8.4

#https://pypi.org/project/python-dotenv/
python-dotenv==0.21.0

#https://pypi.org/project/google-cloud-storage/
google-cloud-storage==2.10.0

#https://pypi.org/project/pillow/
pillow==11.1.0

#https://pypi.org/project/google-auth/
google-auth==2.38.0

#https://pypi.org/project/google-generativeai/
google-genai==0.8.0
